Explicitly cast HANDLE to UINT, otherwise it'll get converted to a far pointer when passed to sprintf()!
[git-p4: depot-paths = "//depot/openwatcom/": change = 24897]

+++ bld/sdk/heapwalk/hwlocal.c
@@ -182,7 +182,7 @@ BOOL FormatLocalHeapListItem( char *line, unsigned index ) {
     }
     sprintf( line, "%04X    %04X  %6d %-9s %4d  %-20s",
              item->wAddress,
-             item->hHandle,
+             (UINT)item->hHandle,
              item->wSize,
              flags,
              item->wcLock,
